complete(fix): move Share on Twitter outside modal
